refactor: rename zip to zip file client
